Initial work for round notes.
Doesn't open keyboard on phones, since the recycler is in a dialog. Either need to make it a new full screen fragment, or let users swap between taking notes and score in the normal adapter. The latter of those options may not be ideal with many players in a game, as the columns will be pretty thin.