Set the score holder to empty when value is 0 (reflects calculator)
This makes it easier to enter scores quickly. Since a score is defaulted to 0, entering a new score would require deleting the 0. This reflects the change to support this in the calculator, by setting the ui edit text to empty when the value is 0. If any other value is in the field the selection starts at the end of the text.